+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - Moved templates to dedicated repository
 - Added `angular` template for pilets
+- Added `svelte` template for pilets